requirements.txt
torch>=1.9.0
torchvision>=0.10.0
faiss-cpu>=1.7.2
matplotlib>=3.4.0
pillow>=8.0.0
tqdm>=4.62.0
requests>=2.25.0
pandas>=1.3.0